compact-block: split out crate from chain
