+++ app/Cargo.toml
@@ -13,6 +13,7 @@ penumbra-shielded-pool = { path = "../crates/shielded-pool" }
 penumbra-sct = { path = "../crates/sct" }
 penumbra-dao = { path = "../crates/dao" }
 penumbra-ibc = { path = "../crates/ibc" }
+penumbra-compact-block = { path = "../crates/compact-block" }
 penumbra-proto = { path = "../proto", features = ["penumbra-storage"] }
 penumbra-transaction = { path = "../transaction" }
 penumbra-storage = { path = "../storage" }
