+++ chain/src/state_key.rs
@@ -47,10 +47,3 @@ pub fn epoch_change_at_height(height: u64) -> String {
 pub fn end_epoch_early() -> &'static str {
     "end_epoch_early"
 }
-
-pub fn stub_compact_block() -> &'static str {
-    "compactblock/stub/compact_block"
-}
-pub fn compact_block(height: u64) -> String {
-    format!("compactblock/compact_block/{height}")
-}
