--- chain/src/sync.rs
@@ -1,5 +1,0 @@
-mod compact_block;
-mod state_payload;
-
-pub use compact_block::CompactBlock;
-pub use state_payload::{StatePayload, StatePayloadDebugKind};
