+++ crates/compact-block/src/component.rs
@@ -1,0 +1,3 @@
+mod view;
+
+pub use view::{StateReadExt, StateWriteExt};
