+++ crates/compact-block/src/event.rs
@@ -1,0 +1,1 @@
+
