+++ crates/shielded-pool/src/event.rs
@@ -1,4 +1,3 @@
-use penumbra_chain::StatePayload;
 use penumbra_crypto::Nullifier;
 use tendermint::abci::{Event, EventAttributeIndexExt};
 
@@ -6,9 +5,11 @@ pub fn spend(nullifier: &Nullifier) -> Event {
     Event::new("spend", [("nullifier", nullifier.to_string()).index()])
 }
 
+/*
 pub fn state_payload(payload: &StatePayload) -> Event {
     Event::new(
         "state_payload",
         [("commitment", payload.commitment().to_string()).index()],
     )
 }
+ */
