+++ pcli/src/command/query/shielded_pool.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use colored_json::prelude::*;
-use penumbra_chain::{CompactBlock, NoteSource, SpendInfo};
+use penumbra_chain::{NoteSource, SpendInfo};
+use penumbra_compact_block::CompactBlock;
 use penumbra_crypto::Nullifier;
 use penumbra_proto::DomainType;
 use penumbra_tct::Commitment;
@@ -40,14 +41,14 @@ pub enum ShieldedPool {
 
 impl ShieldedPool {
     pub fn key(&self) -> String {
-        use penumbra_chain::state_key as chain_state_key;
+        use penumbra_compact_block::state_key as cb_state_key;
         use penumbra_sct::state_key as sct_state_key;
         use penumbra_shielded_pool::state_key;
         match self {
             ShieldedPool::Anchor { height } => sct_state_key::anchor_by_height(*height),
             ShieldedPool::BlockAnchor { height } => sct_state_key::block_anchor_by_height(*height),
             ShieldedPool::EpochAnchor { epoch } => sct_state_key::epoch_anchor_by_index(*epoch),
-            ShieldedPool::CompactBlock { height } => chain_state_key::compact_block(*height),
+            ShieldedPool::CompactBlock { height } => cb_state_key::compact_block(*height),
             ShieldedPool::Commitment { commitment } => state_key::note_source(commitment),
             ShieldedPool::Nullifier { nullifier } => state_key::spent_nullifier_lookup(nullifier),
         }
