+++ view/Cargo.toml
@@ -28,6 +28,7 @@ penumbra-shielded-pool = { path = "../crates/shielded-pool", default-features = 
 penumbra-ibc = { path = "../crates/ibc", default-features = false }
 penumbra-dao = { path = "../crates/dao", default-features = false }
 penumbra-sct = { path = "../crates/sct", default-features = false }
+penumbra-compact-block = { path = "../crates/compact-block", default-features = false }
 # TODO: replace by a penumbra-app
 penumbra-app = { path = "../app" }
